first upgrade

+++ bib.h
@@ -2,8 +2,8 @@
 #define bib_H 
 #include <iostream>
 
-extern int add(std::string firstarg, std::string secondarg);
-extern int sub(std::string firstarg, std::string secondarg);
+extern float add(std::string firstarg, std::string secondarg);
+extern float sub(std::string firstarg, std::string secondarg);
 extern int multi(std::string firstarg, std::string secondarg);
 
 #endif // также предотвращает повторное включение (конец включения если не определена)
